+++ integration-tests/zombie-tests/src/main.spec.ts
@@ -22,7 +22,7 @@ import {
 import { getSmProvider } from "polkadot-api/sm-provider"
 import { getWsProvider } from "polkadot-api/ws-provider/node"
 import { createClient as createRawClient } from "@polkadot-api/substrate-client"
-import { MultiAddress, roc } from "@polkadot-api/descriptors"
+import { getMetadata, MultiAddress, roc } from "@polkadot-api/descriptors"
 import { accounts } from "./keyring"
 import { getPolkadotSigner } from "polkadot-api/signer"
 import { fromHex } from "@polkadot-api/utils"
@@ -67,6 +67,7 @@ describe("E2E", async () => {
   } else {
     client = createClient(
       withPolkadotSdkCompat(getWsProvider("ws://127.0.0.1:9934")),
+      { getMetadata },
     )
   }
   console.log("client started")